+++ app/components/indicators/LoadingMoreTransactions/LoadingMoreTransactions.jsx
@@ -1,6 +1,6 @@
-import StakeyBounceXs from "./StakeyBounce/StakeyBounceExtraSmall";
+import StakeyBounceXs from "../StakeyBounce/StakeyBounceExtraSmall";
 import { FormattedMessage as T } from "react-intl";
-import styles from "./indicators.module.css";
+import styles from "./LoadingMoreTransactions.module.css";
 
 const LoadingMoreTransactionsIndicator = () => (
   <div className={styles.loadingMoreTransactionsIndicator}>
