+++ app/components/indicators/LoadingMoreTransactions/LoadingMoreTransactions.module.css
@@ -1,0 +1,8 @@
+.loadingMoreTransactionsIndicator {
+  padding: 17px 18px 18px;
+  border-radius: 5px;
+  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.2);
+  display: inline-block;
+  margin-top: 2em;
+  background-color: var(--background-back-color);
+}
