+++ app/components/indicators/indicators.module.css
@@ -29,17 +29,6 @@
   font-weight: 700;
 }
 
-/* LoadingMoreTransactions */
-
-.loadingMoreTransactionsIndicator {
-  padding: 17px 18px 18px;
-  border-radius: 5px;
-  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.2);
-  display: inline-block;
-  margin-top: 2em;
-  background-color: var(--background-back-color);
-}
-
 /* NoStats, NoTransactions, ... */
 .noTransactionsIndicator,
 .noTicketsIndicator,
